E: move SelectRelated command to submodule
Moved to a new package `commands`, which will contain files
with the individual classes.

In the workbench initialization, it is sufficient with importing
`envis.commands` to load the commands defined inside.
All 'Commands' are graphical in nature, so the GUI is supposed
to be active already, so we don't need to guard the importing
of `FreeCADGui`.